Move some classes into packages
