+++ src/main/java/duke/Parser.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.command.*;
+
 import java.util.Arrays;
 
 /**
