+++ src/main/java/duke/command/CommandExit.java
@@ -1,4 +1,7 @@
-package duke;
+package duke.command;
+
+import duke.TaskList;
+import duke.ui.UserInterface;
 
 public class CommandExit extends Command {
