+++ src/main/java/duke/ui/UserInterface.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 
 import java.util.Scanner;
